requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.9.1
google-generativeai==0.8.3
google-cloud-texttospeech==2.18.0
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.12
websockets==12.0
